fix(frontend): split timeline page for Next.js page types
Move TimelinePageContent to its own module so page.tsx only default-exports.
Fix TimelineTaskCreateFlow report approver branch by reading user from useAuth.

Made-with: Cursor
